+++ niryo_robot_hardware_stack/ttl_driver/include/ttl_driver/fake_ttl_data.hpp
@@ -43,7 +43,7 @@ class FakeTtlData
             uint32_t max_position{0};
 
             uint32_t temperature{0};
-            uint32_t voltage{0};
+            double voltage{0};
             std::string  firmware{};
         };
