+++ niryo_robot_hardware_stack/ttl_driver/src/abstract_stepper_driver.cpp
@@ -73,7 +73,7 @@ int AbstractStepperDriver::writeSingleCmd(const std::shared_ptr<common::model::A
         case EStepperCommandType::CMD_TYPE_CALIBRATION:
             return startHoming(cmd->getId());
         case EStepperCommandType::CMD_TYPE_CALIBRATION_DIRECTION:
-            return writeHomingDirection(cmd->getId(), cmd->getParam());
+            return writeHomingDirection(cmd->getId(), static_cast<uint8_t>(cmd->getParam()));
         case EStepperCommandType::CMD_TYPE_PING:
             return ping(cmd->getId());
         case EStepperCommandType::CMD_TYPE_CONVEYOR:
@@ -88,7 +88,7 @@ int AbstractStepperDriver::writeSingleCmd(const std::shared_ptr<common::model::A
                 // convert direction and speed into signed speed
                 int8_t dir = static_cast<int8_t>(cmd->getParams().at(2));
                 // normal warning : we need to put an int32 inside an uint32_t
-                uint32_t speed = static_cast<uint32_t>(cmd->getParams().at(1) * dir);
+                uint32_t speed = static_cast<uint32_t>(static_cast<int>(cmd->getParams().at(1)) * dir);
                 return setGoalVelocity(cmd->getId(), speed);
             }
         }
